requirements.txt
slack_bolt
supabase
openai
python-dotenv